+++ src/librender/integrator.cpp
@@ -169,7 +169,7 @@ MTS_VARIANT bool SamplingIntegrator<Float, Spectrum>::render(Scene *scene, Senso
 
     if (!m_stop)
         Log(Info, "Rendering finished. (took %s)",
-            util::time_string(m_render_timer.value(), true));
+            util::time_string((float) m_render_timer.value(), true));
 
     return !m_stop;
 }
